// web/src/lib/api/types/Prefs.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteImages } from "./RemoteImages";
import type { Theme } from "./Theme";

export type Prefs = { page_size: number, conversations: boolean, 
/**
 * Start new messages in the rich-text editor.
 */
compose_html: boolean, theme: Theme, remote_images: RemoteImages, };